UI: initial layout.
Added initial layout for the UI. The UI now consists of a header and footer, a
left and right sidebar and a content area. All but the content have fixed
dimensions. The content is liquid, meaning it fills the entire space that it has
available.

When the window resizes, the content area resizes. The calculations keep count
of things like padding and borders, but ignore margins (since there are none
currently).