UI: The left sidebar now has two resizable divs.
The left sidebar is now split into a market tree div and an items div. Between
the two is a resize handler. When the handler is moved (constrained to the
parent and on the y axis), the two divs will resize accordingly.

Also, updated layout.js to contain code for adjusting columns and the left
sidebar. Fixed some naming issues as well.