UI: Fix empty market group name visual bug.
Market groups that have a blank name were causing a visual bug in which the
groups below them were offset to the right. This was caused by the containing
div not having a set height, thus being 0. The divs, however, contain a floating
fixed height element which was causing this layout flow bug.