UI: Added items search.
The market browser now has an input at the top where you can search for items.
The search function is triggered after you have stopped typing for more than
SEARCH_DELAY miliseconds. A request is then sent to the server which renders the
items template and returns it.

Important: The inv_items table must only contain items from the market groups
that will be displayed in the market browser. This is to save time by not having
to check each item's parent group.