UI: Performance improvement for resize handler.
When the window is being resized, only send a request to update the resize
handler's position if its position has been changed. In adjustLeftSidebar() this
only happens when the window has been resized, so that distance != 0. This
change also prevents extra queries when distance % 2 != 0, because the handler's
position is modified by half of the window's height change.