UI: Include items in the market tree.
buildMarketTree() now has an optional parameter, includeItems, which specifies
whether to include the items in a group. This will be used in the ships tab to
display both market groups and the ships in them. Added new tests.

Changed some Item fields to be nullable so writing tests is easier.