UI: Added fitting window and fit creation.
Another huge one. Even huger. This is the first implementation of the fitting
window. When using the ship browser and clicking the 'Create new fit' link, a
fitting wheel is loaded in the main content pane. For now, the wheel will
highlight the available slots on the ship.

When creating a new fit, a Fit object is saved to the database. The fit will be
tied to the currently logged in user or to the anonymous user's session. Users
can only view their own fits or public ones.

Viewing fits will change the current URL without refreshing the page (using the
new HTML5 features).

Each fit will reside in its own tab. If you click to view a fit which is already
open it will switch to its tab. Closing all the tabs will change the URL to
root.

Moved some functions to ui/utils.py.

Created the Fit and User models.

Turned on the setting to save sessions on every request.

Moved some views from ui to service (fits related).