UI: Added defense stats widget.
Another big one, this is the first stats widget, the defense one. This shows the
resists and EHP for shield, armor and hull. The humanize app has been installed
to help with formatting integers.

To implement this, a new view was added called renderFit that will return the
HTML for the fit and the HTML for the stats. Whenever a fit needs to be
displayed this view will be called.

Each fit tab will be associated with a stats tab. Switching between tabs also
switches between stats. The fits framework has been modified to support this.

Stats widgets are now persistent and synchronized. Whenever a change occurs in
one of the stats tabs, the change will be applied to all the other tabs.
Meanwhile, the positions and states of the widgets is stored in a cookie that is
read when the fit is rendered. Thus, new fits will conform with the current
sorting order.

The jQuery cookie plugin has been set to automatically store cookies in JSON
format.