Refactored buildMarketTree.
It's now a class called MarketTree that takes a MarketGroup instances and
provides a build() method that works just like the previous method. It also
exposes a number of constants used in the serialization of the tree. These
constants need to be unpacked and passed to the template when rendering.

Updated the tests to use the new method and moved them into a folder.